requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.0
Flask-Bcrypt==1.0.1
PyMySQL==1.0.2
Werkzeug==2.3.7
gunicorn==20.1.0
cryptography
twilio==8.10.0
python-dotenv==1.0.0
